requirements.txt
requests>=2.31.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
redis>=5.0.7
SQLAlchemy>=2.0.32
psycopg2-binary>=2.9.9
playwright>=1.46.0
aiohttp>=3.9.0